Privatize ProxyDrawable, expose only IsProxy/HasProxy
