+++ osu.Framework.Tests/Visual/TestCaseIsMaskedAway.cs
@@ -119,7 +119,7 @@ namespace osu.Framework.Tests.Visual
         {
             var box = new Box();
 
-            ProxyDrawable proxy;
+            Drawable proxy;
             Children = new Drawable[]
             {
                 new Container
@@ -151,7 +151,7 @@ namespace osu.Framework.Tests.Visual
         {
             var box = new Box { Position = new Vector2(-1) };
 
-            ProxyDrawable proxy;
+            Drawable proxy;
             Children = new Drawable[]
             {
                 new Container
@@ -182,7 +182,7 @@ namespace osu.Framework.Tests.Visual
         {
             var box = new Box();
 
-            ProxyDrawable proxy;
+            Drawable proxy;
             Children = new Drawable[]
             {
                 new Container
@@ -219,7 +219,7 @@ namespace osu.Framework.Tests.Visual
         {
             var box = new Box { Position = new Vector2(-1) };
 
-            ProxyDrawable proxy;
+            Drawable proxy;
             Children = new Drawable[]
             {
                 new Container
